Rename classes and resources for readability
