+++ src/main/resources/net/gleske/scmfilter/impl/trait/Messages.properties
@@ -23,7 +23,7 @@
 # THE SOFTWARE.
 #
 WildcardSCMFilterTrait.DisplayName=Filter by branch, tag or pull request (with wildcards)
-RegexSCMFilterBranchTrait.DisplayName=Filter by branch (with regular expression)
-RegexSCMFilterTagTrait.DisplayName=Filter by tag (with regular expression)
-RegexSCMFilterPullRequestDestinationTrait.DisplayName=Filter by pull request destination branch (with regular expression)
-RegexSCMFilterPullRequestOriginTrait_DisplayName=Filter by pull request origin branch (with regular expression)
+RegexSCMBranchFilterTrait.DisplayName=Filter by branch (with regular expression)
+RegexSCMTagFilterTrait.DisplayName=Filter by tag (with regular expression)
+RegexSCMPRDestinationFilterTrait.DisplayName=Filter by pull request destination branch (with regular expression)
+RegexSCMPROriginFilterTrait_DisplayName=Filter by pull request origin branch (with regular expression)
