+++ src/main/resources/net/gleske/scmfilter/impl/trait/RegexSCMBranchFilterTrait/config.jelly
