+++ src/main/resources/net/gleske/scmfilter/impl/trait/RegexSCMBranchFilterTrait/help-branchRegex.html
