+++ src/main/resources/net/gleske/scmfilter/impl/trait/RegexSCMPRDestinationFilterTrait/config.jelly
