+++ src/main/resources/net/gleske/scmfilter/impl/trait/RegexSCMPROriginFilterTrait/config.jelly
